+++ Quantcast-iOS-Measurement/Optional/QuantcastMeasurement+Networks.h
@@ -43,6 +43,12 @@
  */
 @interface QuantcastMeasurement (Networks)
 
+/*!
+ @property networkLabels
+ @abstract Property that contains a static network labels
+ @discussion This property can be set to either an NSString or an NSArray of NSStrings.  When set, the label(s) will be automatically passed to all calls which take network labels.  This is a convience property for applications that segment their audience by a fairly static group of labels.   This property can be changed at any time.
+ */
+@property (retain,nonatomic) id<NSObject> networkLabels;
 
 /*!
  @method beginMeasurementSessionWithAPIKey:attributedNetwork:userIdentifier:appLabels:networkLabels:appIsDirectedAtChildren: