+++ Quantcast-iOS-Measurement/QuantcastParameters.h
@@ -105,8 +105,8 @@
 // Quantcast Measurement SDK
 //
 
-#define QCMEASUREMENT_API_VERSION               @"1_2_12"
-#define QCMEASUREMENT_API_IDENTIFIER            @"iOS_1.2.12"
+#define QCMEASUREMENT_API_VERSION               @"1_2_13"
+#define QCMEASUREMENT_API_IDENTIFIER            @"iOS_1.2.13"
 #define QCMEASUREMENT_CONN_TIMEOUT_SECONDS      60
 
 #ifndef QCMEASUREMENT_UPLOAD_URL